Split gwemopt.utils into subscripts
